clean up code
